+++ tools/winlibs.R
@@ -4,7 +4,9 @@ if(getRversion() < "3.3.0") {
 
 # For details see: https://github.com/rwinlib/gdal3
 VERSION <- commandArgs(TRUE)
-if(!file.exists(sprintf("../windows/gdal3-%s/include/gdal/gdal.h", VERSION))){
+# wrong path: if(!file.exists(sprintf("../windows/gdal3-%s/include/gdal/gdal.h", VERSION))){
+testfile <- sprintf("../windows/gdal3-%s/include/gdal-%s/gdal.h", VERSION, VERSION)
+if(!file.exists(testfile)){  
   if(getRversion() < "3.3.0") setInternet2()
   download.file(sprintf("https://github.com/rwinlib/gdal3/archive/v%s.zip", VERSION), "lib.zip", quiet = TRUE)
   dir.create("../windows", showWarnings = FALSE)