+++ NEWS.md
@@ -4,7 +4,7 @@
 
 * `st_write` writes non-spatial tables when given a plain `data.frame` or `tbl_df`; #1345
 
-* the default for `stringsAsFactors` in `st_read` and `st_sf` is FALSE for R version > 4.1.0
+* the default for `stringsAsFactors` in `st_read` and `st_sf` is `FALSE` for R version >= 4.1.0
 
 # version 0.9-3
 
@@ -12,7 +12,7 @@
 
 * Windows CRAN binaries use GDAL 3.0.4, PROJ 6.3.1 and GEOS 3.8.0, thanks to Jeroen Ooms' rwinlib work; #1275
 
-* `plot.sf` gains an `extent` argument to set the extent (xlim, ylim) of the plot; `extent` must be an object with an `st_bbox()` method, such as an `sf` object; #1193
+* `plot.sf` gains an `extent` argument to set the extent (xlim, ylim) of the plot; `extent` must be an object with an `st_bbox()` method, such as an `sf` or a `stars` object; #1193
 
 # version 0.9-2